pyproject.toml
[project]
name = "djkhaled"
version = "0.1.0"
description = "A music bot for Discord"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py[voice]==2.5.2",
    "pydantic==2.11.4",
    "pytimeparse>=1.1.8",
    "yt-dlp>=2025.4.30",
]

[dependency-groups]
dev = [
    "mypy==1.15.0",
    "ruff==0.11.10",
]

[project.scripts]
djkhaled = "djkhaled.__main__:app"

[tool.mypy]
strict = true
pretty = true
enable_error_code = ["ignore-without-code"]

[tool.ruff]
line-length = 120
